test(development-harness): move ordinal property test into existing suite
The prior commit added a new plugins/development-harness/tests/test_ordinal_mapper.py,
duplicating the basename of the canonical suite at
plugins/development-harness/backlog_core/tests/unit/test_ordinal_mapper.py — a pytest
module-name collision hazard of the same class AGENTS.md documents for conftest.

The generator/validator agreement property (#2969 Finding 2) is not covered by the
canonical suite, so it is relocated there rather than dropped, following the
TestSectionRoundTripProperty hypothesis convention in test_github_sync.py.
